fix: prevent gift_mapper overflow with u64 saturating_mul
Problem: apply_rule multiplied rule.strength_add (u8) * gift.num (u32)
as i16 on both sides. When gift.num is large (e.g. 999), 255*999
overflows i16, causing a debug panic.

Fix: Use u64 saturating_mul for raw_delta (255 * 2^32 won't overflow).
StrengthChangeEvent.delta is widened to i32. handle_gift uses i32
saturating_add + clamp to the effective limit, preserving arithmetic
precision until the final device output. Gift log display uses the
u64 raw_delta directly.

+++ src/engine/types.rs
@@ -15,7 +15,7 @@ pub struct StrengthStatus {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct StrengthChangeEvent {
     pub channel: Channel,
-    pub delta: i16,
+    pub delta: i32,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub absolute: Option<u8>,
     pub source: StrengthSource,
